requirements.txt
streamlit
firebase-admin
requests
pandas
python-dotenv
twilio
matplotlib
google-generativeai
reportlab
streamlit-option-menu
plotly